Add support for background workers in the browser. (#1)
Uses an abstraction over Web Workers and worker_threads. A bit of a pain
due to transferList.

It requires a bit of a dance to create the Workers, particularly as
Safari has no nested worker support. Browser clients will need updating
to the new incantation.

Also adds minimal file system operations. These are a stop-gap solution.